IR: minor, move IrMemberAccessExpression utils to a separate file
To help git understand that IrMemberAccessExpression is moved in the
subsequent commits where it's changed to be auto-generated.
